+++ internal/app/rolloutStrategy/dto_adaptive.go
@@ -14,7 +14,7 @@ type rsAdaptivePhaseDto struct {
 func (r rsAdaptivePhaseDto) validate() error {
 	return validation.ValidateStruct(&r,
 		validation.Field(&r.MinFeedback, validation.Required, validation.Min(int64(1))),
-		validation.Field(&r.MaxStepPct, validation.Required, validation.Min(float64(1)), validation.Max(float64(100))),
+		validation.Field(&r.MaxStepPct, validation.Required, validation.Min(1.0), validation.Max(100.0)),
 		validation.Field(&r.IntervalMins, validation.Required, validation.Min(int64(1))),
 	)
 }
